+++ app/delete_product_handler.go
@@ -8,14 +8,13 @@ import (
 	"github.com/go-chi/chi"
 )
 
-// DeleteProductHandler create a hanlder to delete a product resource.
+// DeleteProductHandler create a handler to delete a product resource.
 func (a *App) DeleteProductHandler() http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
 		sku := chi.URLParam(r, "sku")
-
-		err := a.Service.DeleteProduct(r.Context(), sku)
-		if err != nil {
+		if err := a.Service.DeleteProduct(r.Context(), sku); err != nil {
 			fmt.Fprintf(os.Stderr, "delete product sku=%q failed: %v", sku, err)
+			w.WriteHeader(http.StatusInternalServerError) // 500 Internal Server Error
 			return
 		}
 		w.Header().Del("Content-Type")