+++ app/get_image.go
@@ -1,0 +1,30 @@
+package app
+
+import (
+	"database/sql"
+	"encoding/json"
+	"fmt"
+	"net/http"
+	"os"
+
+	"github.com/go-chi/chi"
+)
+
+// GetImageHandler returns a handler function that gets a single image by UUID.
+func (a *App) GetImageHandler() http.HandlerFunc {
+	return func(w http.ResponseWriter, r *http.Request) {
+		uuid := chi.URLParam(r, "uuid")
+		product, err := a.Service.GetImage(r.Context(), uuid)
+		if err != nil {
+			if err == sql.ErrNoRows {
+				w.WriteHeader(http.StatusNotFound)
+				return
+			}
+			fmt.Fprintf(os.Stderr, "service: GetImage(ctx, %q) error: %+v", uuid, err)
+			w.WriteHeader(http.StatusInternalServerError) // 500 Internal Server Error
+			return
+		}
+		w.WriteHeader(http.StatusOK) // 200 OK
+		json.NewEncoder(w).Encode(*product)
+	}
+}