+++ model/postgres/postgres.go
@@ -83,8 +83,9 @@ type ProductData struct {
 	Spec    string `json:"specification"`
 }
 
+// ProductImage struct holds a row of the product_images table.
 type ProductImage struct {
-	ID        uint
+	id        uint
 	ProductID uint
 	UUID      string
 	SKU       string
@@ -1324,7 +1325,7 @@ func (m *PgModel) CreateImageEntry(ctx context.Context, c *CreateProductImage) (
 		c.W, c.H, c.Path, c.Typ,
 		c.Ori,
 		c.Pri, c.Size, c.Q,
-		c.GSURL).Scan(&p.ID, &p.ProductID, &p.UUID, &p.SKU, &p.W, &p.H, &p.Path, &p.Typ, &p.Ori, &p.Up, &p.Pri, &p.Size, &p.Q, &p.GSURL, &p.Data, &p.Created, &p.Modified)
+		c.GSURL).Scan(&p.id, &p.ProductID, &p.UUID, &p.SKU, &p.W, &p.H, &p.Path, &p.Typ, &p.Ori, &p.Up, &p.Pri, &p.Size, &p.Q, &p.GSURL, &p.Data, &p.Created, &p.Modified)
 	if err != nil {
 		return nil, err
 	}
@@ -1332,11 +1333,57 @@ func (m *PgModel) CreateImageEntry(ctx context.Context, c *CreateProductImage) (
 	return &p, nil
 }
 
-// GetImageEntries returns a list of all images associated to a given product SKU.
-func (m *PgModel) GetImageEntries(ctx context.Context, sku string) ([]*ProductImage, error) {
+// ImagePathExists return true if the image with the given path exists in
+// the database.
+func (m *PgModel) ImagePathExists(ctx context.Context, path string) (bool, error) {
+	query := `SELECT id FROM product_images WHERE path = $1`
+	var id int
+	if err := m.db.QueryRowContext(ctx, query, path).Scan(&id); err != nil {
+		if err == sql.ErrNoRows {
+			return false, nil
+		}
+		return false, errors.Wrapf(err, "query row context path=%q query=%q", path, query)
+	}
+	return true, nil
+}
+
+
+// ImageUUIDExists return true if the image with the given UUID exists in
+// the database.
+func (m *PgModel) ImageUUIDExists(ctx context.Context, uuid string) (bool, error) {
+	query := `SELECT id FROM product_images WHERE uuid = $1`
+	var id int
+	if err := m.db.QueryRowContext(ctx, query, uuid).Scan(&id); err != nil {
+		if err == sql.ErrNoRows {
+			return false, nil
+		}
+		return false, errors.Wrapf(err, "query row context uuid=%q query=%q", uuid, query)
+	}
+	return true, nil
+}
+
+// GetProductImageByUUID returns a ProductImage by the given UUID.
+// If the product is not found returns nil, error indicating not found.
+func (m *PgModel) GetProductImageByUUID(ctx context.Context, uuid string) (*ProductImage, error) {
 	query := `
 		SELECT id, product_id, uuid, sku, w, h, path, typ, ori, up, pri, size, q,
 		gsurl, data, created, modified
+		FROM product_images
+		WHERE uuid = $1
+	`
+	p := ProductImage{}
+	if err := m.db.QueryRowContext(ctx, query, uuid).Scan(&p.id, &p.ProductID, &p.UUID, &p.SKU, &p.W, &p.H, &p.Path, &p.Typ, &p.Ori, &p.Up, &p.Pri, &p.Size, &p.Q, &p.GSURL, &p.Data, &p.Created, &p.Modified); err != nil {
+		return nil, err
+	}
+	return &p, nil
+}
+
+// GetImageBySKU returns a list of all images associated to a given product SKU.
+func (m *PgModel) GetImageBySKU(ctx context.Context, sku string) ([]*ProductImage, error) {
+	query := `
+		SELECT
+			id, product_id, uuid, sku, w, h, path, typ, ori, up, pri, size, q,
+			gsurl, data, created, modified
 		FROM product_images
 		WHERE sku = $1
 		ORDER BY pri ASC
@@ -1346,17 +1393,15 @@ func (m *PgModel) GetImageEntries(ctx context.Context, sku string) ([]*ProductIm
 		return nil, err
 	}
 	defer rows.Close()
-
 	images := make([]*ProductImage, 0, 16)
 	for rows.Next() {
 		p := ProductImage{}
-		err = rows.Scan(&p.ID, &p.ProductID, &p.UUID, &p.SKU, &p.W, &p.H, &p.Path, &p.Typ, &p.Ori, &p.Up, &p.Pri, &p.Size, &p.Q, &p.GSURL, &p.Data, &p.Created, &p.Modified)
+		err = rows.Scan(&p.id, &p.ProductID, &p.UUID, &p.SKU, &p.W, &p.H, &p.Path, &p.Typ, &p.Ori, &p.Up, &p.Pri, &p.Size, &p.Q, &p.GSURL, &p.Data, &p.Created, &p.Modified)
 		if err != nil {
 			return nil, err
 		}
 		images = append(images, &p)
 	}
-
 	if err = rows.Err(); err != nil {
 		return nil, err
 	}
@@ -1374,21 +1419,39 @@ func (m *PgModel) ConfirmImageUploaded(ctx context.Context, uuid string) (*Produ
 		gsurl, data, created, modified
 	`
 	p := ProductImage{}
-	err := m.db.QueryRowContext(ctx, query, uuid).Scan(&p.ID, &p.ProductID, &p.UUID, &p.SKU, &p.W, &p.H, &p.Path, &p.Typ, &p.Ori, &p.Up, &p.Pri, &p.Size, &p.Q, &p.GSURL, &p.Data, &p.Created, &p.Modified)
+	err := m.db.QueryRowContext(ctx, query, uuid).Scan(&p.id, &p.ProductID, &p.UUID, &p.SKU, &p.W, &p.H, &p.Path, &p.Typ, &p.Ori, &p.Up, &p.Pri, &p.Size, &p.Q, &p.GSURL, &p.Data, &p.Created, &p.Modified)
 	if err != nil {
 		return nil, err
 	}
 	return &p, nil
 }
 
-// DeleteImageEntry deletes an image entry row from the product_images
+// DeleteProductImage deletes an image entry row from the product_images
 // table by UUID.
-func (m *PgModel) DeleteImageEntry(ctx context.Context, uuid string) (int64, error) {
+func (m *PgModel) DeleteProductImage(ctx context.Context, uuid string) (int64, error) {
 	query := `
 		DELETE FROM product_images
 		WHERE uuid = $1
 	`
 	res, err := m.db.ExecContext(ctx, query, uuid)
+	if err != nil {
+		return -1, err
+	}
+	count, err := res.RowsAffected()
+	if err != nil {
+		return -1, err
+	}
+	return count, nil
+}
+
+// DeleteAllProductImages Images deletes all images from the product_images table
+//  associated to the product with the given SKU.
+func (m *PgModel) DeleteAllProductImages(ctx context.Context, sku string) (int64, error) {
+	query := `
+		DELETE FROM product_images
+		WHERE sku = $1
+	`
+	res, err := m.db.ExecContext(ctx, query, sku)
 	if err != nil {
 		return -1, err
 	}